add branch

+++ .github/scripts/deploy.sh
@@ -6,16 +6,23 @@ nvm use 20
 
 temp_dir=$(mktemp -d)
 
-git clone git@github.com:Raid-Hub/RaidHub-API.git "$temp_dir"
+git clone -b $1 git@github.com:Raid-Hub/RaidHub-API.git "$temp_dir"
+if [ $? -ne 0 ]; then
+    exit 1
+fi
+
 cd $temp_dir
 
 bun install --frozen-lockfile
 bun prisma generate
-bun compile $2
+bun run compile $2
 
 mv $temp_dir/$2 ~/../RaidHub/API/$2
-rm -rf "$temp_dir"
+if [ $? -ne 0 ]; then
+    exit 1
+fi
 
+rm -rf "$temp_dir"
 if [ $? -ne 0 ]; then
     exit 1
 fi
